C4-1019: Ingestion to snovault migration
