+++ snovault/project/loadxl.py
@@ -1,0 +1,6 @@
+# Loadxl related functions which may be overriden by an implementing app,
+# e.g. Foursight or CGAP portal, using the dcicutils project_utils mechanism.
+
+class SnovaultProjectLoadxl:
+    def loadxl_order(self):
+        return []
